Extract functionality to `utils`
