+++ crates/lang/ir/src/ir/trait_def/config.rs
@@ -15,7 +15,7 @@
 use crate::{
     ast,
     error::ExtError as _,
-    ir::config::WhitelistedAttributes,
+    utils::WhitelistedAttributes,
 };
 use syn::spanned::Spanned;
